Tidy up common code
