+++ src/functionalTest/groovy/org/openmicroscopy/tasks/FileGeneratorTaskTest.groovy
@@ -1,0 +1,16 @@
+package org.openmicroscopy.tasks
+
+import org.openmicroscopy.AbstractBaseTest
+
+class FileGeneratorTaskTest extends AbstractBaseTest {
+
+    def "can create task"() {
+        buildFile << """
+            task genSingleFile(type: )
+        
+        
+        """
+    }
+
+
+}
